source/requirements.txt
beautifulsoup4==4.11.1
markdownify==0.14.1
numpy==2.0.0
Pillow==10.4.0
python-dateutil==2.8.1
PyYAML==6.0.1
qrcode==6.1
requests==2.32.2
Unidecode==1.3.4
discord.py==2.5.2
click
